+++ CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Application Server pub/sub integrations race condition during shutdown.
 - Console webhook templates empty headers error.
 - Console MQTT URL validation.
+- AFCntDown from the application-layer is respected when skipping application payload crypto.
 
 ### Security